[FRAGMENTS] - Migrate main activities to fragmetns
