+++ app/build.gradle.kts
@@ -59,6 +59,7 @@ dependencies {
     implementation(libs.commons.codec)
     implementation(libs.androidx.security.crypto)
 
+    implementation(libs.androidx.fragment.ktx)
     implementation(libs.logging.interceptor)
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
