+++ app/src/main/AndroidManifest.xml
@@ -16,10 +16,7 @@
         android:theme="@style/Theme.EldarWallet"
         tools:targetApi="31">
         <activity
-            android:name=".ui.activities.contactlesspay.ContactlessPayActivity"
-            android:exported="false" />
-        <activity
-            android:name=".ui.activities.qrpay.QrPayActivity"
+            android:name=".ui.ApplicationActivity"
             android:exported="false" />
         <activity
             android:name=".ui.activities.addcard.AddDigitalCardActivity"
@@ -33,9 +30,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".ui.activities.main.MainActivity"
-            android:exported="true" />
     </application>
 
 </manifest>
