+++ app/src/main/java/com/org/marton/studio/project/eldarwallet/ui/ApplicationActivity.kt
@@ -1,97 +1,80 @@
-package com.org.marton.studio.project.eldarwallet.ui.activities.main
+package com.org.marton.studio.project.eldarwallet.ui
 
 import android.content.Intent
-import android.os.Build
 import android.os.Bundle
-import android.se.omapi.Session
 import android.view.Menu
 import android.view.MenuItem
-import android.widget.TextView
 import androidx.activity.enableEdgeToEdge
-import androidx.activity.viewModels
-import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
+import androidx.fragment.app.commit
+import androidx.fragment.app.replace
 import com.google.android.material.bottomnavigation.BottomNavigationView
-import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.org.marton.studio.project.eldarwallet.R
-import com.org.marton.studio.project.eldarwallet.ui.activities.addcard.AddDigitalCardActivity
-import com.org.marton.studio.project.eldarwallet.ui.activities.contactlesspay.ContactlessPayActivity
 import com.org.marton.studio.project.eldarwallet.ui.activities.login.LoginActivity
-import com.org.marton.studio.project.eldarwallet.ui.activities.main.adapter.DigitalCardAdapter
-import com.org.marton.studio.project.eldarwallet.ui.activities.qrpay.QrPayActivity
+import com.org.marton.studio.project.eldarwallet.ui.fragments.contactlesspay.ContactlessPayFragment
+import com.org.marton.studio.project.eldarwallet.ui.fragments.main.MainFragment
+import com.org.marton.studio.project.eldarwallet.ui.fragments.qrpay.QrPayFragment
 import com.org.marton.studio.project.eldarwallet.utils.SessionData
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
+class ApplicationActivity : AppCompatActivity() {
 
-    private val viewModel: MainViewModel by viewModels()
     companion object {
         var selectedItemId = R.id.main_activity_tab
     }
-    @RequiresApi(Build.VERSION_CODES.P)
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_main)
+        setContentView(R.layout.activity_application)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
             insets
         }
+
         val toolbar = findViewById<Toolbar>(R.id.toolbar)
         setSupportActionBar(toolbar)
 
-        val addCardButton: FloatingActionButton = findViewById(R.id.agregarTarjetaButton)
         val bottomNavigationView: BottomNavigationView = findViewById(R.id.bottomNavigationView)
-        val balanceTextView: TextView = findViewById(R.id.balanceTextView)
-        val usernameTextView: TextView = findViewById(R.id.usernameEditText)
-
-
         bottomNavigationView.setOnItemSelectedListener { item ->
             selectedItemId = item.itemId
-
             when (item.itemId) {
                 R.id.qr_paid_tab -> {
-                    intent = Intent(this, QrPayActivity::class.java)
-                    startActivity(intent)
+                    supportFragmentManager.commit {
+                        replace<QrPayFragment>(R.id.frameApplicationContainer)
+                        setReorderingAllowed(true)
+                        addToBackStack("replacement")
+                    }
                     true
                 }
 
                 R.id.main_activity_tab -> {
+                    supportFragmentManager.commit {
+                        replace<MainFragment>(R.id.frameApplicationContainer)
+                        setReorderingAllowed(true)
+                        addToBackStack("replacement")
+                    }
                     true
                 }
 
                 R.id.contacless_paid_tab -> {
-                    intent = Intent(this, ContactlessPayActivity::class.java)
-                    startActivity(intent)
+                    supportFragmentManager.commit {
+                        replace<ContactlessPayFragment>(R.id.frameApplicationContainer)
+                        setReorderingAllowed(true)
+                        addToBackStack("replacement")
+                    }
                     true
                 }
+
                 else -> false
             }
-
         }
         bottomNavigationView.selectedItemId = selectedItemId
-
-        viewModel.userData.observe(this) { userData ->
-            usernameTextView.text = getGreattings(userData.userName + " " + userData.userLastname)
-            balanceTextView.text = String.format("$ %.2f", userData.balance)
-
-            val recyclerView: RecyclerView = findViewById(R.id.tarjetasRecyclerView)
-            val adapter = DigitalCardAdapter(userData.cards ?: emptyList())
-            recyclerView.adapter = adapter
-            recyclerView.layoutManager = LinearLayoutManager(this)
-        }
-
-        addCardButton.setOnClickListener {
-            intent = Intent(this, AddDigitalCardActivity::class.java)
-            startActivity(intent)
-        }
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
@@ -109,9 +92,5 @@ class MainActivity : AppCompatActivity() {
             }
             else -> super.onOptionsItemSelected(item)
         }
-    }
-
-    private fun getGreattings(s: String): String {
-        return "Hi $s!"
     }
 }
