+++ app/src/main/res/values/strings.xml
@@ -30,5 +30,7 @@
     <string name="lastname">Lastname</string>
     <string name="logout">Logout</string>
     <string name="pay">Pay</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 
 </resources>
