+++ gradle/libs.versions.toml
@@ -2,6 +2,7 @@
 agp = "8.5.0"
 androidDatabaseSqlcipher = "4.5.0"
 commonsCodec = "1.15"
+fragmentKtx = "1.8.3"
 hilt = "2.51.1"
 kotlin = "1.9.0"
 coreKtx = "1.13.1"
@@ -24,6 +25,7 @@ tinkAndroid = "1.8.0"
 [libraries]
 android-database-sqlcipher = { module = "net.zetetic:android-database-sqlcipher", version.ref = "androidDatabaseSqlcipher" }
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
+androidx-fragment-ktx = { module = "androidx.fragment:fragment-ktx", version.ref = "fragmentKtx" }
 androidx-security-crypto = { module = "androidx.security:security-crypto", version.ref = "securityCrypto" }
 commons-codec = { module = "commons-codec:commons-codec", version.ref = "commonsCodec" }
 junit = { group = "junit", name = "junit", version.ref = "junit" }
